+++ internal/align/provider_test.go
@@ -43,8 +43,26 @@ func TestProviderAttributeOrder(t *testing.T) {
 	got := string(file.Bytes())
 	exp := `provider "aws" {
   alias   = "west"
+  profile = "default"
+  region  = "us-east-1"
+}`
+	require.Equal(t, exp, got)
+}
+
+func TestProviderAttributeOrderNoAlias(t *testing.T) {
+	src := []byte(`provider "aws" {
   region  = "us-east-1"
   profile = "default"
+  version = "~> 4.0"
+}`)
+	file, diags := hclwrite.ParseConfig(src, "in.tf", hcl.InitialPos)
+	require.False(t, diags.HasErrors())
+	require.NoError(t, alignpkg.Apply(file, &alignpkg.Options{}))
+	got := string(file.Bytes())
+	exp := `provider "aws" {
+  profile = "default"
+  region  = "us-east-1"
+  version = "~> 4.0"
 }`
 	require.Equal(t, exp, got)
 }